refactor: move core.oauth to verticals.auth
