+++ ggshield/cmd/auth/login.py
@@ -8,8 +8,8 @@ from ggshield.core.client import create_client
 from ggshield.core.config import Config
 from ggshield.core.constants import DEFAULT_INSTANCE_URL
 from ggshield.core.errors import UnexpectedError
-from ggshield.core.oauth import OAuthClient
 from ggshield.core.utils import clean_url
+from ggshield.verticals.auth.oauth import OAuthClient
 
 
 def validate_login_path(
